Add validation for frames on the control stream
Motivation:

Add basic validation for frames on the control stream and also correctly validate for push promise frames

Modifications:

- Add validation for control stream
- Refactor code to easier share between server and client (once client support is added)
- Move inner classes so its easier to test things etc

Result:

Better validation
