+++ src/main/java/io/netty/incubator/codec/http3/QpackStreamHandler.java
@@ -1,0 +1,68 @@
+/*
+ * Copyright 2020 The Netty Project
+ *
+ * The Netty Project licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+package io.netty.incubator.codec.http3;
+
+import io.netty.channel.ChannelHandlerContext;
+import io.netty.channel.ChannelInboundHandlerAdapter;
+import io.netty.channel.socket.ChannelInputShutdownEvent;
+import io.netty.incubator.codec.quic.QuicStreamChannel;
+import io.netty.util.ReferenceCountUtil;
+
+final class QpackStreamHandler extends ChannelInboundHandlerAdapter {
+    static final QpackStreamHandler INSTANCE = new QpackStreamHandler();
+
+    private QpackStreamHandler() {
+    }
+
+    @Override
+    public boolean isSharable() {
+        return true;
+    }
+
+    @Override
+    public void channelRead(ChannelHandlerContext ctx, Object msg) {
+        ReferenceCountUtil.release(msg);
+    }
+
+    @Override
+    public void channelReadComplete(ChannelHandlerContext ctx) {
+        ctx.fireChannelReadComplete();
+
+        // QPACK streams should always be processed, no matter what the user is doing in terms of configuration
+        // and AUTO_READ.
+        Http3CodecUtils.readIfNoAutoRead(ctx);
+    }
+
+    @Override
+    public void userEventTriggered(ChannelHandlerContext ctx, Object evt) {
+        if (evt instanceof ChannelInputShutdownEvent) {
+            criticalStreamClosed(ctx);
+        }
+        ctx.fireUserEventTriggered(evt);
+    }
+
+    @Override
+    public void channelInactive(ChannelHandlerContext ctx) {
+        criticalStreamClosed(ctx);
+        ctx.fireChannelInactive();
+    }
+
+    // See https://www.ietf.org/archive/id/draft-ietf-quic-qpack-19.html#section-4.2
+    private void criticalStreamClosed(ChannelHandlerContext ctx) {
+        Http3CodecUtils.closeParent((QuicStreamChannel) ctx.channel(),
+                Http3ErrorCode.H3_CLOSED_CRITICAL_STREAM, "Critical QPACK stream closed.");
+    }
+}
